Project structure refactoring, Add Comments, Store config refactor
