+++ optaweb-tsp-ui/src/components/Location.tsx
