+++ optaweb-tsp-ui/src/components/TspMap.test.tsx
